evals: scrub stale Ray comments + collapse LevanterTpuEvaluator base
Second-pass cleanup after the launch_evaluate_with_ray removal:

- Drop stale Ray-framed comment in lm_evaluation_harness_evaluator's
  tokenizer-staging helper and docstring in evalchemy's _run_evalchemy_in_process
  (the worker-death-kills-subprocess reasoning still applies, just not
  Ray-specifically).
- Inline LevanterTpuEvaluator.model_name_or_path into LevanterLmEvalEvaluator
  and delete the intermediate base class. With launch_evaluate_with_ray gone
  it was a 17-line file wrapping a single static helper and one subclass.
